+++ clean-code-katas/clean-code-katas.sln
@@ -25,6 +25,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "todictionary", "todictionar
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "fromromannumerals.tests", "function-katas\tests\fromromannumerals.tests\fromromannumerals.tests.csproj", "{04F1F976-3A5A-42C0-8B48-FEAA66138BB4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "fromromannumerals", "function-katas\fromromannumerals\fromromannumerals.csproj", "{D6851572-558A-483B-90F6-8DA9AF03408B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -43,6 +45,10 @@ Global
 		{04F1F976-3A5A-42C0-8B48-FEAA66138BB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{04F1F976-3A5A-42C0-8B48-FEAA66138BB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{04F1F976-3A5A-42C0-8B48-FEAA66138BB4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D6851572-558A-483B-90F6-8DA9AF03408B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D6851572-558A-483B-90F6-8DA9AF03408B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D6851572-558A-483B-90F6-8DA9AF03408B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D6851572-558A-483B-90F6-8DA9AF03408B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -52,6 +58,7 @@ Global
 		{B4F85FD8-A834-4A40-B55F-4887D0AC8F94} = {D06B579B-4270-40D4-AE72-481828C5C9DC}
 		{8BEDB266-9DCA-4BE6-B3C3-152095E71D69} = {A6B4F3FA-686E-44FC-B765-E35A7010999D}
 		{04F1F976-3A5A-42C0-8B48-FEAA66138BB4} = {D06B579B-4270-40D4-AE72-481828C5C9DC}
+		{D6851572-558A-483B-90F6-8DA9AF03408B} = {A6B4F3FA-686E-44FC-B765-E35A7010999D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {67C6EEA4-00F5-49D9-B1C1-68D845BC7AFD}
