+++ clean-code-katas/function-katas/tests/fromromannumerals.tests/fromromannumerals.tests.csproj
@@ -59,6 +59,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="RomanToDecimalConverter.cs" />
     <Compile Include="RomanToDecimalConverterTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
